# Conflicts:
#	common/translation-app/backend/package.json

+++ packages/sdk/src/app/(forms)/_hooks/useFormsActions.ts
@@ -86,6 +86,8 @@ export default function useFormsActions({ t }: UseFormsActionsProps) {
 
   const deleteFromList = useCallback(
     async (fileId: number) => {
+      if (!window.confirm(t("Common:DeletePermanently") + "?")) return;
+
       try {
         await deleteFile(fileId, false, true);
         const newItems = formsListStore.items.filter((f) => f.id !== fileId);
@@ -136,6 +138,8 @@ export default function useFormsActions({ t }: UseFormsActionsProps) {
 
   const deleteFolderFromList = useCallback(
     async (folderId: number) => {
+      if (!window.confirm(t("Common:DeletePermanently") + "?")) return;
+
       try {
         await deleteFolderApi(folderId, false, true);
         const newFolders = formsListStore.folders.filter(
@@ -153,6 +157,7 @@ export default function useFormsActions({ t }: UseFormsActionsProps) {
     async (file: TFile) => {
       try {
         await manageFormFilling(file.id, FormFillingManageAction.Start);
+        toastr.success(t("Common:ReadyToFillOut"));
         await fetchSection();
       } catch (error) {
         toastr.error(error as string);
